% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urp_top_cors.R
\name{top_cors}
\alias{top_cors}
\title{Find highesy correlations in your data}
\usage{
top_cors(data)
}
\arguments{
\item{data}{Your data in the form of a matrix, data.frame, or data.table}
}
\value{
A data.frame of all the pairwise correlations of your data, in order of decreasing magnitude
}
\description{
Find highesy correlations in your data
}
\examples{
top_cors(df1)
}
